backend/requirements.txt
fastapi
uvicorn
python-multipart
rapidocr-onnxruntime
Pillow
